Move history to its own component
